Allow parsing of context to renderers (#8)
In this PR we replace the (optional) passing of individual block/span token class lists to the renderers, and instead allow the entire `ParseContext` to be passed. Passing an existing `ParseContext` means that it will be not be reset, which is required if we wish to perform nested renderers, within the same global context (not resetting gathered link/footnote definitions). It also (slightly) simplifies the API.

Additionally, we add more documentation about using the package *via* python, as an API.